\name{convert.samples.c4photo}
\alias{convert.samples.c4photo}
\title{Convert samples for c4photo}
\usage{
  convert.samples.c4photo(trait.samples)
}
\arguments{
  \item{trait.samples}{a matrix or dataframe of samples
  from the trait distribution}
}
\value{
  matrix or dataframe with values transformed
}
\description{
  convert parameters from PEcAn database default units to
  c4photo defaults
}
\details{
  Performs model specific unit conversions on a a list of
  trait values, such as those provided to write.config
}
\author{
  David LeBauer
}

