\name{model2netcdf.c4photo}
\alias{model2netcdf.c4photo}
\title{Function to convert c4photo model output to standard netCDF format}
\usage{
  model2netcdf.c4photo(outdir, run.id)
}
\arguments{
  \item{outdir}{Location of c4photo model output}

  \item{run.id}{Name of c4photo model output file.}
}
\description{
  Function to convert c4photo model output to standard
  netCDF format
}
\author{
  Shawn Serbin, Michael Dietze, David LeBauer
}

