\name{remove.config.c4photo}
\alias{remove.config.c4photo}
\title{Clear out previous c4photo config and parameter files.}
\usage{
  remove.config.c4photo(main.outdir, settings)
}
\arguments{
  \item{main.outdir}{Primary PEcAn output directory (will
  be depreciated)}

  \item{settings}{PEcAn settings file}
}
\value{
  nothing, removes config files as side effect
}
\description{
  Clear out previous config and parameter files.
}
\author{
  Shawn Serbin, David LeBauer
}

