\name{start.runs.c4photo}
\alias{start.runs.c4photo}
\title{Start c4photo model runs}
\usage{
  start.run.c4photo(run.id)

  start.runs.c4photo()
}
\value{
  nothing, starts run as side effect
}
\description{
  Start c4photo model runs on local or remote server

  Function to start all runs of c4photo model in directory
}
\author{
  Michael Dietze, David LeBauer, Shawn Serbin, Carl
  Davidson

  David LeBauer
}

