\name{write.config.c4photo}
\alias{write.config.c4photo}
\title{Write configuration files for the c4photo model}
\usage{
  write.config.c4photo(defaults = NULL, trait.values,
    settings, outdir, run.id)
}
\arguments{
  \item{defaults}{named list with default model parameter
  values}

  \item{trait.values}{}

  \item{settings}{}

  \item{outdir}{}

  \item{run.id}{}
}
\value{
  nothing, writes configuration file as side effect
}
\description{
  Writes a configuration files for the c4photo model
}
\author{
  David LeBauer
}

